feat: email verification
