+++ dhis-2/dhis-support/dhis-support-db-migration/src/main/resources/org/hisp/dhis/db/migration/2.42/V2_42_14__Add_verifed_email_to_userinfo.sql
@@ -1,3 +1,3 @@
 ALTER TABLE userinfo ADD COLUMN if not exists verifiedemail VARCHAR(255);
-ALTER TABLE userinfo ADD COLUMN if not exists verifiedemailtoken VARCHAR(255);
+ALTER TABLE userinfo ADD COLUMN if not exists emailverificationtoken VARCHAR(255);
 ALTER TABLE userinfo ADD CONSTRAINT userinfo_verifiedemail_key UNIQUE (verifiedemail);
