+++ dhis-2/dhis-test-web-api/src/test/java/org/hisp/dhis/webapi/controller/AccountControllerTest.java
@@ -28,21 +28,29 @@
 package org.hisp.dhis.webapi.controller;
 
 import static org.hisp.dhis.test.utils.Assertions.assertContainsOnly;
+import static org.hisp.dhis.test.web.WebClientUtils.assertStatus;
 import static org.hisp.dhis.test.webapi.Assertions.assertWebMessage;
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.util.List;
 import java.util.Set;
 import org.hisp.dhis.jsontree.JsonList;
 import org.hisp.dhis.jsontree.JsonMixed;
 import org.hisp.dhis.jsontree.JsonObject;
+import org.hisp.dhis.message.MessageSender;
+import org.hisp.dhis.outboundmessage.OutboundMessage;
 import org.hisp.dhis.setting.SettingKey;
 import org.hisp.dhis.setting.SystemSettingManager;
+import org.hisp.dhis.test.message.FakeMessageSender;
 import org.hisp.dhis.test.web.HttpStatus;
 import org.hisp.dhis.test.webapi.PostgresControllerIntegrationTestBase;
 import org.hisp.dhis.user.User;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.transaction.annotation.Transactional;
 
 /**
  * Tests the {@link AccountController} using (mocked) REST requests.
@@ -51,6 +59,7 @@ import org.springframework.beans.factory.annotation.Autowired;
  */
 class AccountControllerTest extends PostgresControllerIntegrationTestBase {
   @Autowired private SystemSettingManager systemSettingManager;
+  @Autowired private MessageSender messageSender;
 
   @Test
   void testRecoverAccount_NotEnabled() {
@@ -215,6 +224,124 @@ class AccountControllerTest extends PostgresControllerIntegrationTestBase {
     JsonMixed response = GET("/account/linkedAccounts").content(HttpStatus.OK);
     JsonList<JsonObject> users = response.getList("users", JsonObject.class);
     assertEquals(3, users.size());
+  }
+
+  @Test
+  @Transactional
+  void testVerifyEmailWithTokenTwice() {
+    User user = switchToNewUser("kent");
+
+    String emailAddress = user.getEmail();
+    assertStatus(HttpStatus.CREATED, POST("/account/sendEmailVerification"));
+    OutboundMessage emailMessage = assertMessageSendTo(emailAddress);
+    String token = extractTokenFromEmailText(emailMessage.getText());
+    assertNotNull(token);
+
+    assertStatus(HttpStatus.OK, GET("/account/verifyEmail?token=" + token));
+    user = userService.getUser(user.getId());
+    assertTrue(userService.isEmailVerified(user));
+
+    assertStatus(HttpStatus.CONFLICT, GET("/account/verifyEmail?token=" + token));
+  }
+
+  @Test
+  void testSendEmailVerification() {
+    User user = switchToNewUser("clark");
+
+    String emailAddress = user.getEmail();
+    assertStatus(HttpStatus.CREATED, POST("/account/sendEmailVerification"));
+    OutboundMessage emailMessage = assertMessageSendTo(emailAddress);
+    String token = extractTokenFromEmailText(emailMessage.getText());
+    assertNotNull(token);
+
+    user = userService.getUser(user.getId());
+    assertFalse(userService.isEmailVerified(user));
+  }
+
+  @Test
+  void testVerifyEmailWithToken() {
+    User user = switchToNewUser("lex");
+
+    String emailAddress = user.getEmail();
+    assertStatus(HttpStatus.CREATED, POST("/account/sendEmailVerification"));
+    OutboundMessage emailMessage = assertMessageSendTo(emailAddress);
+    String token = extractTokenFromEmailText(emailMessage.getText());
+    assertValidEmailVerificationToken(token);
+
+    assertStatus(HttpStatus.OK, GET("/account/verifyEmail?token=" + token));
+    user = userService.getUser(user.getId());
+    assertTrue(userService.isEmailVerified(user));
+  }
+
+  @Test
+  void testVerifyWithBadToken() {
+    switchToNewUser("zod");
+    assertWebMessage(
+        "Conflict",
+        409,
+        "ERROR",
+        "Verification token is invalid",
+        GET("/account/verifyEmail?token=eviltoken").content(HttpStatus.CONFLICT));
+  }
+
+  @Test
+  void testSendEmailWithoutEmail() {
+    User user = switchToNewUser("metallo");
+    user.setEmail(null);
+    assertWebMessage(
+        "Conflict",
+        409,
+        "ERROR",
+        "Email is not set",
+        POST("/account/sendEmailVerification").content(HttpStatus.CONFLICT));
+  }
+
+  @Test
+  void testSendEmailIsAlreadyVerifiedBySameUser() {
+    User user = switchToNewUser("brainiac");
+    user.setVerifiedEmail(user.getEmail());
+    userService.updateUser(user);
+    assertWebMessage(
+        "Conflict",
+        409,
+        "ERROR",
+        "Email is already verified",
+        POST("/account/sendEmailVerification").content(HttpStatus.CONFLICT));
+  }
+
+  @Test
+  void testSendEmailIsAlreadyVerifiedByAnotherUser() {
+    User user = switchToNewUser("doomsday");
+
+    user.setVerifiedEmail(user.getEmail());
+    userService.updateUser(user);
+
+    User anotherUser = switchToNewUser("ultraman");
+    anotherUser.setEmail(user.getEmail());
+
+    assertWebMessage(
+        "Conflict",
+        409,
+        "ERROR",
+        "Email is already in use by another account",
+        POST("/account/sendEmailVerification").content(HttpStatus.CONFLICT));
+  }
+
+  private void assertValidEmailVerificationToken(String token) {
+    User user = userService.getUserByVerificationToken(token);
+    assertNotNull(user);
+  }
+
+  private OutboundMessage assertMessageSendTo(String email) {
+    List<OutboundMessage> messagesByEmail =
+        ((FakeMessageSender) messageSender).getMessagesByEmail(email);
+    assertFalse(messagesByEmail.isEmpty());
+    return messagesByEmail.get(0);
+  }
+
+  private String extractTokenFromEmailText(String message) {
+    int tokenPos = message.indexOf("?token=");
+    return message.substring(tokenPos + 7, message.indexOf('\n', tokenPos)).trim();
   }
 
   private static void assertMessage(String key, String value, String message, JsonMixed response) {
