+++ dhis-2/dhis-web-api/src/main/java/org/hisp/dhis/webapi/controller/AccountController.java
@@ -56,6 +56,7 @@ import org.hisp.dhis.common.OpenApi;
 import org.hisp.dhis.configuration.ConfigurationService;
 import org.hisp.dhis.dxf2.webmessage.WebMessage;
 import org.hisp.dhis.dxf2.webmessage.WebMessageException;
+import org.hisp.dhis.feedback.ConflictException;
 import org.hisp.dhis.feedback.ErrorCode;
 import org.hisp.dhis.organisationunit.OrganisationUnit;
 import org.hisp.dhis.security.PasswordManager;
@@ -78,6 +79,7 @@ import org.hisp.dhis.user.UserService;
 import org.hisp.dhis.webapi.mvc.annotation.ApiVersion;
 import org.hisp.dhis.webapi.utils.HttpServletRequestPaths;
 import org.hisp.dhis.webapi.webdomain.user.UserLookups;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
@@ -90,6 +92,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.ResponseStatus;
 
 /**
  * @author Lars Helge Overland
@@ -521,6 +524,43 @@ public class AccountController {
   public ResponseEntity<Map<String, String>> validatePasswordPost(
       @RequestParam String password, HttpServletResponse response) {
     return ResponseEntity.ok().cacheControl(noStore()).body(validatePassword(password));
+  }
+
+  @PostMapping("/sendEmailVerification")
+  @ResponseStatus(HttpStatus.CREATED)
+  public void sendEmailVerification(@CurrentUser User currentUser, HttpServletRequest request)
+      throws ConflictException {
+    if (Strings.isNullOrEmpty(currentUser.getEmail())) {
+      throw new ConflictException("Email is not set");
+    }
+    if (userService.isEmailVerified(currentUser)) {
+      throw new ConflictException("Email is already verified");
+    }
+    if (userService.getUserByVerifiedEmail(currentUser.getEmail()) != null) {
+      throw new ConflictException("Email is already in use by another account");
+    }
+
+    // Generate a new email verification token and send it, we do this in two steps:
+    // 1. Generate and save the token to the user
+    // 2. Send the token to the user's email
+    // This is because email delivery is unreliable can fail/respond false even if email is sent,
+    // and true if email is not sent/received.
+    String token = userService.generateAndSetNewEmailVerificationToken(currentUser);
+    boolean successfullySent =
+        userService.sendEmailVerificationToken(
+            currentUser, token, HttpServletRequestPaths.getContextPath(request));
+
+    if (!successfullySent) {
+      throw new ConflictException("Failed to send email verification token");
+    }
+  }
+
+  @GetMapping("/verifyEmail")
+  @ResponseStatus(HttpStatus.OK)
+  public void verifyEmail(@RequestParam String token) throws ConflictException {
+    if (!userService.verifyEmail(token)) {
+      throw new ConflictException("Verification token is invalid");
+    }
   }
 
   // ---------------------------------------------------------------------
